# perf: small improvements in Synth