--- a/ci-benchmark/benchmark.graphql
+++ b/ci-benchmark/benchmark.graphql
@@ -1,9 +1,9 @@
-schema @server(port: 8000, hostname: "0.0.0.0") @upstream(baseURL: "http://jsonplaceholder.typicode.com") {
+schema @server(port: 8000, hostname: "0.0.0.0") {
   query: Query
 }
 
 type Query @cache(maxAge: 30000) {
-  posts: [Post] @http(path: "/posts")
+  posts: [Post] @http(url: "http://jsonplaceholder.typicode.com/posts")
 }
 
 type User {
@@ -20,5 +20,5 @@ type Post {
   userId: Int!
   title: String!
   body: String!
-  user: User @http(path: "/users/{{.value.userId}}")
+  user: User @http(url: "http://jsonplaceholder.typicode.com/users/{{.value.userId}}")
 }