--- a/ci-benchmark/nginx-benchmark.graphql
+++ b/ci-benchmark/nginx-benchmark.graphql
@@ -1,7 +1,6 @@
 schema
   @server(hostname: "0.0.0.0", port: 8000)
   @upstream(
-    baseURL: "http://jsonplaceholder.typicode.com"
     poolMaxIdlePerHost: 200
     tcpKeepAlive: 60
     proxy: {url: "http://127.0.0.1:3000"}
@@ -10,7 +9,7 @@ schema
 }
 
 type Query {
-  posts: [Post] @http(path: "/posts")
+  posts: [Post] @http(url: "http://jsonplaceholder.typicode.com/posts")
 }
 
 type User {
@@ -27,5 +26,5 @@ type Post {
   userId: Int!
   title: String!
   body: String!
-  user: User @http(path: "/users/{{value.userId}}")
+  user: User @http(url: "http://jsonplaceholder.typicode.com/users/{{value.userId}}")
 }