--- a/ci-benchmark/nginx-benchmark.graphql
+++ b/ci-benchmark/nginx-benchmark.graphql
@@ -1,10 +1,6 @@
 schema
   @server(hostname: "0.0.0.0", port: 8000)
-  @upstream(
-    poolMaxIdlePerHost: 200
-    tcpKeepAlive: 60
-    proxy: {url: "http://127.0.0.1:3000"}
-  ) {
+  @upstream(poolMaxIdlePerHost: 200, tcpKeepAlive: 60, proxy: {url: "http://127.0.0.1:3000"}) {
   query: Query
 }