@meskill 2024-10-22
RPS changes locally:

## posts title

### before

```
Running 30s test @ http://localhost:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.95ms   14.36ms 486.95ms   99.11%
    Req/Sec    13.96k     2.02k   18.78k    89.16%
  1670008 requests in 27.10s, 8.37GB read
  Socket errors: connect 0, read 0, write 0, timeout 100
Requests/sec:  61623.45
Transfer/sec:    316.29MB
```

### after

```
Running 30s test @ http://localhost:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.89ms   15.14ms 551.80ms   99.09%
    Req/Sec    15.33k     2.66k   27.94k    85.63%
  1832428 requests in 27.66s, 9.18GB read
  Socket errors: connect 0, read 0, write 0, timeout 100
Requests/sec:  66243.75
Transfer/sec:    340.01MB
```

## big query

### before

```
Running 30s test @ http://localhost:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    10.03ms    4.46ms 117.46ms   85.96%
    Req/Sec     2.59k     0.87k   16.07k    93.05%
  304120 requests in 29.64s, 15.04GB read
  Socket errors: connect 0, read 0, write 0, timeout 100
Requests/sec:  10259.24
Transfer/sec:    519.61MB
```

### after

```
Running 30s test @ http://localhost:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.26ms    3.05ms  68.16ms   75.51%
    Req/Sec     2.73k   236.40     3.05k    83.42%
  325544 requests in 30.05s, 16.10GB read
Requests/sec:  10832.93
Transfer/sec:    548.66MB
```